refactor: extract error handling from AgentWorkspace
